fix: Fix the BookComponent display porblem
why?
    -- 發現 最下面的 BookComponent 會被 切掉一小段，之前沒發現，已修正
    -- 修正 common.js 的參數錯置

+++ src/components/BookComponent.vue
@@ -1,24 +1,23 @@
 <template>
   <div class="bookDiv">
-    <b-card :title="title" class="bookCardButton" @click="toBookDetailPage">
+    <b-card class="bookCardButton" @click="toBookDetailPage">
       <span class="material-symbols-outlined deleteIcon" @click="handleDelete">
         delete
       </span>
+      <template #header>
+        <h4 class="mb-0">{{ title }}</h4>
+      </template>
       <b-card-img :src="img" top style="width: 60%"></b-card-img>
       <b-card-body>
-        <b-card-text>
-          {{ author }}
-        </b-card-text>
-        <b-card-text>
-          {{ description }}
-        </b-card-text>
+        <b-card-text> 作者: {{ author }} </b-card-text>
+        <b-card-text> {{ description }} </b-card-text>
       </b-card-body>
     </b-card>
   </div>
 </template>
 
 <script>
-import { ref, reactive, onMounted } from "vue";
+import { ref, onMounted } from "vue";
 import { useRouter } from "vue-router";
 
 import { common } from "../lib/common.js";
@@ -26,12 +25,12 @@ import { common } from "../lib/common.js";
 export default {
   name: "BookComponent",
   props: ["book"],
-  setup(props) {
+  setup(props, { emit }) {
     const router = useRouter();
 
     const id = ref(props.book.id);
-    const title = reactive(props.book.title);
-    const author = reactive(props.book.author);
+    const title = ref(props.book.title);
+    const author = ref(props.book.author);
     const description = ref(props.book.description);
     const img = ref(props.book.image);
 
@@ -42,9 +41,13 @@ export default {
     }
 
     async function handleDelete() {
-      const result = await common.deleteBookById(id.value);
-      if (typeof result === "object") {
-        alert("刪除成功");
+      const ok = confirm("確定刪除?");
+      if (ok) {
+        const result = await common.deleteBookById(id.value);
+        if (typeof result === "object") {
+          alert("刪除成功");
+          emit("getBooks");
+        }
       }
     }
 
@@ -72,8 +75,7 @@ export default {
 
 <style scoped>
 .bookDiv {
-  width: 20rem;
-  height: 25rem;
+  height: 100%;
   flex: 1 0 45%;
   text-align: center;
   margin: 0.5rem;
@@ -81,7 +83,9 @@ export default {
 
 .bookCardButton {
   width: 20rem;
-  height: 25rem;
+  min-height: 25rem;
+  height: 100%;
+  max-height: 30rem;
   background-color: white;
   margin: 0rem auto;
   flex: 1 0 40%;
@@ -98,11 +102,12 @@ export default {
   float: right;
   border: 1px solid;
   border-radius: 1rem;
-  margin: -3rem -0.9rem 0 0;
+  margin: -0.8rem -0.9rem 0 0;
   z-index: 100;
 }
 
 .deleteIcon:hover {
   background-color: red;
+  transition-duration: 0.5s;
 }
 </style>
